requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
sendgrid==6.11.0
mailgun==0.1.1
python-dotenv==1.0.1
